+++ src/__init__.py
@@ -1,1 +1,7 @@
 """Package: hcp-segmentation-engine"""
+
+from src.rfm_scorer import (  # noqa: F401
+    compute_rfm_scores,
+    get_top_hcps,
+    summarise_rfm_segments,
+)